First files (still not done ;)
The main page started looking good but the dark/white mode is still not good (only the 'section' is changing the color from white to black and from black to white). The button will need new position and new colors for the  whole page. The logo is still with pixels and it will be changed with HD soon.